% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{boot_ci}
\alias{boot_ci}
\title{Calculate empirical bootstrap confidence intervals}
\usage{
boot_ci(x, times, fun = "mean", probs = c(0.1, 0.5, 0.9), w = 1)
}
\arguments{
\item{x}{Numeric vector}

\item{times}{Number of bootstrap samples}

\item{fun}{Function to use}

\item{probs}{Numeric vector of probabilities}

\item{w}{Weight vector}
}
\value{
Numeric vector of confidence intervals
}
\description{
Calculate empirical bootstrap confidence intervals
}
