% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count}
\alias{count}
\title{Alternative to tabulate}
\usage{
count(x, sort = FALSE)
}
\arguments{
\item{x}{Vector}

\item{sort}{Logical, sort in descending order of count?}
}
\value{
List of unique values of x and counts of each unique value
}
\description{
Alternative to tabulate
}
