% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{db_write_files}
\alias{db_write_files}
\title{Write data files to database table}
\usage{
db_write_files(files, file_db, table_name, batch_size = 1)
}
\arguments{
\item{files}{Paths of delimited data files}

\item{file_db}{Path of database}

\item{table_name}{Database table name}

\item{batch_size}{Number of data files to write to database table per batch}
}
\description{
Write delimited data files to database table in batches. Data files must have
the same headers. If the database file already exists, it will be deleted
first.
}
