% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{equal}
\alias{equal}
\title{Test that all elements in vector are equal}
\usage{
equal(x, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{x}{Vector}

\item{tol}{Tolerance}
}
\value{
Logical
}
\description{
Test that all elements in vector are equal
}
