% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{equal_all}
\alias{equal_all}
\title{Test that all elements in a vector are equal}
\usage{
equal_all(x, na.rm = FALSE, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{x}{Vector}

\item{na.rm}{Logical, remove NAs before comparing?}

\item{tol}{Tolerance}
}
\value{
Logical
}
\description{
Test that all elements in a vector are equal
}
