% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance.R
\name{fin_amort_pay}
\alias{fin_amort_pay}
\title{Calculate payment of amortization}
\usage{
fin_amort_pay(principal, rate, n)
}
\arguments{
\item{principal}{Principal starting amount}

\item{rate}{Interest rate per period}

\item{n}{Number of periods}
}
\value{
Numeric vector of payment per period
}
\description{
Calculate payment of amortization
}
