% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance.R
\name{fin_amort_rate}
\alias{fin_amort_rate}
\title{Calculate estimated interest rate of amortization}
\usage{
fin_amort_rate(principal, payment, n, tol = 1e-05)
}
\arguments{
\item{principal}{Principal starting amount}

\item{payment}{Payment per period}

\item{n}{Number of periods}

\item{tol}{Tolerance}
}
\value{
Numeric vector of interest rate per period
}
\description{
Calculate estimated interest rate of amortization
}
