% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance.R
\name{fin_n}
\alias{fin_n}
\title{Calculate number of compounding periods}
\usage{
fin_n(pv, pay, rate)
}
\arguments{
\item{pv}{Present value}

\item{pay}{Payment per period}

\item{rate}{Rate of interest per period}
}
\value{
Number of compounding periods
}
\description{
Calculate number of compounding periods
}
