% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance.R
\name{fin_pv}
\alias{fin_pv}
\title{Calculate present value}
\usage{
fin_pv(pay, rate, n)
}
\arguments{
\item{pay}{Payment per period}

\item{rate}{Rate of interest per period}

\item{n}{Number of compounding periods}
}
\value{
Present value
}
\description{
Calculate present value
}
