% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance.R
\name{fin_rate}
\alias{fin_rate}
\title{Calculate rate of interest per period}
\usage{
fin_rate(pv, pay, n, guess = 0.1, tol = 1e-05, max_iter = 100)
}
\arguments{
\item{pv}{Present value}

\item{pay}{Payment per period}

\item{n}{Number of compounding periods}

\item{guess}{Starting guess for solving rate of interest, default 0.1}

\item{tol}{Tolerance for solution}

\item{max_iter}{Maximum iterations to find solution}
}
\value{
Rate of interest per period
}
\description{
Calculate rate of interest per period
}
