% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadm.extract.R
\name{gadm.extract}
\alias{gadm.extract}
\title{Extract layers by country from GADM GeoPackage file}
\usage{
gadm.extract(input, output, countries = NULL, level = 0:4, ...)
}
\arguments{
\item{input}{GeoPackage file to read from}

\item{output}{name of file to save output to}

\item{countries}{country or countries to limit results to, if \code{NULL} returns
all countries}

\item{level}{level(s) of administrative units 0:4 to extract; note level 4 is
not available for all countries}

\item{...}{additional arguments passed to \code{\link[sf:st_write]{sf::st_write()}}}
}
\description{
Extract one or more levels of administrative unit geometries from
}
\examples{
\dontrun{
gadm.extract("gadm34_levels.gpkg", "Nordics.gpkg",
             c("Denmark", "Finland", "Iceland", "Norway", "Sweden"),
             level = 0:4)
}
}
