% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{identical_all}
\alias{identical_all}
\title{Test objects in list for exact equality}
\usage{
identical_all(x)
}
\arguments{
\item{x}{List of objects}
}
\value{
A single logical value.
}
\description{
Test objects in list for exact equality
}
