% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{map_us_county}
\alias{map_us_county}
\title{Create US county map}
\usage{
map_us_county(data, fill, type = c("polygons", "centroids"),
  size = NULL, state = NULL)
}
\arguments{
\item{data}{Data frame containing columns county_fips, numeric vector fill,
and optionally a numeric vector size}

\item{fill}{Name of numeric vector in data to fill county colors by}

\item{type}{polygons or centroids}

\item{size}{When type equals centroids, name of numeric vector in data to
determine circle size}

\item{state}{Character vector of state FIPS codes to include in map}
}
\description{
Create US county map
}
