% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{map_us_metro}
\alias{map_us_metro}
\title{Create US metro map}
\usage{
map_us_metro(data, fill, size = NULL)
}
\arguments{
\item{data}{Data frame containing columns cbsa_code, numeric vector fill,
and optionally a numeric vector size}

\item{fill}{Name of numeric vector in data to fill county colors by}

\item{size}{Name of numeric vector in data to determine circle size}
}
\description{
Create US metro map
}
