% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmccp.R
\name{mcmccp}
\alias{mcmccp}
\title{Coefficient Plots for MCMC Output}
\usage{
mcmccp(mod, pars, ci, plot = T)
}
\arguments{
\item{mod}{a \code{stanfit} object.}

\item{pars}{a scalar or vector of the parameters you wish to include in the table.
\code{stanfit} objects can use either the individual parameter names, or the
names of the indexed parameter to retrieve the entire parameter e.g.
\code{pars = 'beta'} will return \code{beta[1]}, \code{beta[2]}, and \code{beta[3]} for a Stan
model with a three element beta parameter.}

\item{ci}{a scalar indicating the confidence level of the uncertainty intervals.}

\item{plot}{logical indicating whether to return a \code{ggplot} object or the
underlying tidy DataFrame.}
}
\value{
a \code{ggplot} object or a tidy DataFrame.
}
\description{
Coefficient plots for MCMC output using \code{ggplot2}
}
