% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcme.R
\name{mcmcme}
\alias{mcmcme}
\title{Marginal Effects Plots for MCMC Output}
\usage{
mcmcme(mod, main, int, moderator, point.est = "mean", seq = 100,
  ci = 0.95, plot = T)
}
\arguments{
\item{mod}{a \code{stanfit} object.}

\item{main}{a character with the name of the parameter of interest in the
interaction term.}

\item{int}{a character with the name of the moderating parameter in the
interaction term.}

\item{moderator}{a vector of values that the moderating parameter takes on
in the data.}

\item{point.est}{a character indicating whether to use the mean or median for
point estimates in the plot.}

\item{seq}{a numeric giving the number of moderator values used to generate
the marginal effects plot.}

\item{ci}{a scalar indicating the confidence level of the uncertainty intervals.}

\item{plot}{logical indicating whether to return a \code{ggplot} object or the
underlying tidy DataFrame.}
}
\value{
a \code{ggplot} object or a tidy DataFrame.
}
\description{
Marginal effects plots for MCMC output using \code{ggplot2}
}
