% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointpolydist.R
\name{point.poly.dist}
\alias{point.poly.dist}
\title{Point-Polygon Distances}
\usage{
point.poly.dist(poly, point, max = T)
}
\arguments{
\item{poly}{A simplefeatures object of class polygon or multipolygon.}

\item{point}{A simplefeatures object of class point.}

\item{max}{Logical; return maximum or minimum distance?}
}
\value{
Maximum or minimum distance between a point and a polygon.
}
\description{
Calculate the maximum or minimum possible distance from a point to the edge
of a given polygon.
}
\examples{
library(sf)
polys <- st_sfc(st_polygon(list(rbind(c(0,0), c(0,1), c(1,1), c(1,0), c(0,0)))),
crs = 4326)
points <- st_sfc(st_multipoint(rbind(c(.25, .5), c(.75, .5))), crs = 4326)
point.poly.dist(points, polys)
}
