% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.summary2}
\alias{print.summary2}
\title{Print summary2}
\usage{
\method{print}{summary2}(x, digits = 4, n_rows = 10, ...)
}
\arguments{
\item{x}{Data frame}

\item{digits}{Minimum number of significant digits to use}

\item{n_rows}{Maximum number of rows to print}

\item{...}{Other arguments passed on to print}
}
\description{
Print summary2
}
