% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printing.R
\name{print.tbl_df}
\alias{print.tbl_df}
\title{Print tibble as a data.frame}
\usage{
\method{print}{tbl_df}(x, digits = 4, n_rows = 10)
}
\arguments{
\item{x}{Data frame}

\item{digits}{Minimum number of significant digits to use}

\item{n_rows}{Maximum number of rows to print}
}
\description{
Print tibble as a data.frame
}
