% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{roll_ema}
\alias{roll_ema}
\title{Rolling exponential moving average}
\usage{
roll_ema(x, n, alpha = 0.1, ...)
}
\arguments{
\item{x}{Numeric vector}

\item{n}{Window size}

\item{alpha}{Smoothing parameter}

\item{...}{Addional arguments passed on to \code{roll::roll_mean()}}
}
\value{
Numeric vector of size x
}
\description{
Rolling exponential moving average
}
