% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWmisc.R
\docType{package}
\name{RWmisc-package}
\alias{RWmisc}
\alias{RWmisc-package}
\title{RWmisc: Rob Williams Miscellaneous}
\description{
Contains convenience functions for working with
    spatial data across multiple UTM zones and recoding
    data commonly used in international relations.
}
\details{
Convenience functions for presenting MCMC output of Bayesian estimation and working with spatial data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jayrobwilliams/RWmisc}
  \item Report bugs at \url{https://github.com/jayrobwilliams/RWmisc/issues}
}

}
\keyword{internal}
