% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfFunctions.R
\name{st_drop_geometry}
\alias{st_drop_geometry}
\title{Simplefeature Downgrade}
\usage{
st_drop_geometry(x)
}
\arguments{
\item{x}{A simplefeature object.}
}
\value{
A DataFrame.
}
\description{
Downgrade a Simplefeature Geometry Object to a DataFrame.
}
