% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary2}
\alias{summary2}
\title{Alternative to \code{base::summary()} for data frames}
\usage{
summary2(data, digits = 4)
}
\arguments{
\item{data}{A data frame}

\item{digits}{Number of significant digits to display for mean and quantiles}
}
\value{
A data frame
}
\description{
Alternative to \code{base::summary()} for data frames
}
