% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary2_by}
\alias{summary2_by}
\title{Summary of data frame by a variable}
\usage{
summary2_by(data, by, vars, digits = 4)
}
\arguments{
\item{data}{A data frame}

\item{by}{A length one character vector of a variable in \code{data} to group by}

\item{vars}{Character vector of variables in \code{data} to keep in summary}

\item{digits}{Number of significant digits to display for mean and quantiles}
}
\value{
A data frame
}
\description{
Summary of data frame by a variable
}
