% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.R
\name{tabulate2}
\alias{tabulate2}
\title{Alternative to tabulate}
\usage{
tabulate2(x)
}
\arguments{
\item{x}{Vector}
}
\value{
List of unique values of x and counts of each unique value
}
\description{
Alternative to tabulate
}
