% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_rw}
\alias{theme_rw}
\title{A ggplot2 theme}
\usage{
theme_rw(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{Base font size}

\item{base_family}{Base font family}
}
\description{
A ggplot2 theme
}
