% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfFunctions.R
\name{UTM.functions}
\alias{UTM.functions}
\alias{UTMzones}
\alias{projectUTM}
\title{UTM Convenience Functions}
\usage{
UTMzones(long)

projectUTM(sfo)
}
\arguments{
\item{long}{A vector of longitude values.}

\item{sfo}{A simplefeature object in latitude-longitude CRS.}
}
\value{
UTM vector of zone numbers.

A simplefeature object projected to UTM CRS.
}
\description{
Functions for converting latitude-longitude data to UTM.
}
