% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.R
\name{wtd_quantile}
\alias{wtd_quantile}
\title{Weighted quantiles}
\usage{
wtd_quantile(x, w = NULL, probs = seq(0, 1, 0.25), na.rm = FALSE)
}
\arguments{
\item{x}{Numeric vector}

\item{w}{Numeric vector of weights}

\item{probs}{Numeric vector of probabilities}

\item{na.rm}{Logical}
}
\value{
Numeric vector of length probs
}
\description{
Weighted quantiles
}
