% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance.R
\name{fin_rate}
\alias{fin_rate}
\title{Calculate estimated interest rate of amortization}
\usage{
fin_rate(principal, payment, n, tol = 1e-05)
}
\arguments{
\item{principal}{Principal starting amount, single value}

\item{payment}{Payment per period, single value}

\item{n}{Number of periods, single value}

\item{tol}{Tolerance, single value}
}
\value{
Single value numeric vector of interest rate per period
}
\description{
Calculate estimated interest rate of amortization
}
