% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usage.R
\name{pkg_count_calls}
\alias{pkg_count_calls}
\title{Count function calls of package imports from source}
\usage{
pkg_count_calls(dir)
}
\arguments{
\item{dir}{Directory of package}
}
\value{
Data frame of rows as files in directory \code{R} and columns as function
calls from a package that is imported.
}
\description{
Count function calls of package imports from source
}
