% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectUTM.R
\name{projectUTM}
\alias{projectUTM}
\alias{projectUTM.sf}
\alias{projectUTM.SpatialPointsDataFrame}
\alias{projectUTM.SpatialPolygonsDataFrame}
\title{Project to UTM}
\usage{
projectUTM(x, ...)

\method{projectUTM}{sf}(x)

\method{projectUTM}{SpatialPointsDataFrame}(x)

\method{projectUTM}{SpatialPolygonsDataFrame}(x)
}
\arguments{
\item{x}{An \code{sf} or \code{sp} object in latitude-longitude CRS.}
}
\value{
An \code{sf} or \code{sp} object projected to UTM CRS.
}
\description{
Project an object in latitude/longitude to UTM.
}
\examples{
library(sf)
nc <- st_read(system.file("shape/nc.shp", package="sf"))
st_crs(projectUTM(nc))
}
