% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWmisc.R
\docType{package}
\name{RWmisc-package}
\alias{RWmisc}
\alias{RWmisc-package}
\title{RWmisc: Miscellaneous Spatial Functions}
\description{
Contains convenience functions for working with spatial data across
    multiple UTM zones, raster-vector operations common in the analysis of 
    conflict data, and converting degrees, minutes, and seconds latitude and
    longitude coordinates to decimal degrees.
}
\details{
Convenience functions for working with spatial data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jayrobwilliams/RWmisc}
  \item Report bugs at \url{https://github.com/jayrobwilliams/RWmisc/issues}
}

}
\author{
\strong{Maintainer}: Rob Williams \email{jayrobwilliams@gmail.com} (\href{https://orcid.org/0000-0001-9259-3883}{ORCID})

}
\keyword{internal}
