% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwmisc-package.R
\docType{package}
\name{rwmisc-package}
\alias{rwmisc}
\alias{rwmisc-package}
\title{rwmisc: Ray Wright Miscellaneous}
\description{
Miscellaneous functions.
}
\author{
\strong{Maintainer}: Ray Wright \email{raywright88@gmail.com}

}
\keyword{internal}
