% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary2}
\alias{summary2}
\title{Alternative to summary for data frames}
\usage{
summary2(data, probs = seq(0, 1, 0.25))
}
\arguments{
\item{data}{Data frame}

\item{probs}{Numeric vector of probabilities}
}
\value{
Data frame
}
\description{
Alternative to summary for data frames
}
