% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary2}
\alias{summary2}
\title{Alternative to summary for data frames}
\usage{
summary2(df, probs = seq(0, 1, 0.25), uniques = FALSE)
}
\arguments{
\item{df}{Data frame}

\item{probs}{Numeric vector of probabilities}

\item{uniques}{Logical, count uniques?}
}
\value{
Data frame
}
\description{
Alternative to summary for data frames
}
