% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary2_by}
\alias{summary2_by}
\title{Alternative to summary for data frames, by groups}
\usage{
summary2_by(data, by, vars, probs = seq(0, 1, 0.25))
}
\arguments{
\item{data}{Data frame}

\item{by}{Character vector of variables in \code{data} to group by}

\item{vars}{Character vector of variables in \code{data} to keep in summary}

\item{probs}{Numeric vector of probabilities}
}
\value{
Data frame
}
\description{
Alternative to summary for data frames, by groups
}
