% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary2_by}
\alias{summary2_by}
\title{Alternative to summary for data frames, by groups}
\usage{
summary2_by(data, by, vars)
}
\arguments{
\item{data}{A data frame}

\item{by}{A length one character vector of a variable in \code{data} to group by}

\item{vars}{Character vector of variables in \code{data} to keep in summary}
}
\value{
A data frame
}
\description{
Alternative to summary for data frames, by groups
}
