% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_rw.R
\name{theme_rw}
\alias{theme_rw}
\title{Blank Theme}
\usage{
theme_rw()
}
\value{
A ggplot \link[ggplot2:theme]{theme} object.
}
\description{
A ggplot theme with no grid elements or gray background.
}
\examples{
ggplot2::ggplot(mtcars, ggplot2::aes(x = hp, y = mpg)) +
ggplot2::geom_point() +
theme_rw()
}
