% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTMfunctions.R
\name{UTM.functions}
\alias{UTM.functions}
\alias{long2UTM}
\alias{UTMzones}
\alias{chooseUTM}
\title{UTM Convenience Functions}
\usage{
long2UTM(long)

UTMzones(long)

chooseUTM(long)
}
\arguments{
\item{long}{A vector of longitude values.}
}
\value{
UTM vector of zone numbers.

UTM vector of zone numbers.

UTM zone number.
}
\description{
Functions for converting latitude-longitude data to UTM.
}
\examples{
long2UTM(c(-90, 0, 90))
UTMzones(c(-90, 90, 90))
chooseUTM(c(-90, -80, -70))
}
